feat: complete v3.0.0 security cleanup + pipeline docs + coverage to 97.9%
Security fixes (P3-1, P3-2):
- FluxAssembler: rebuild Modifier-first chain in Instantiate to prevent
  silent wrong results in Release when VFF deserializes Modifier chains
- Remove InternalsVisibleTo(FluxFormula.Editor) — replaced only internal
  access (formula.Type debug log) with public API

Tests (+31: 301→332):
- FileFluxFileFormatterTests (12 new) — Save/Load/round-trip/extension
- ModifierFormulaTests (+17 Modifier + 3 Modifier-first chain regression)
- FluxEvaluatorTests: IFluxDefinition.GetOperatorName DIM
- Fix 2 TOper→byte migration assertions in ArithmeticTests/DataStructsTests

Docs (12 new pipeline articles, 中英双语):
- lexer, compiler, instruction, evaluator, jit, platform
- Fix TOper→TDef in overview.md (both CN/EN)
- Add sidebar entries in VitePress config
- Session summary: docs/technical/session-2026-06-25.md

Coverage: 97.2% → 97.9% line, FileFluxFileFormatter 0%→100%,
FluxModifier 32.4%→89.1%, IFluxDefinition 0%→100%

+++ tests/FluxFormula.Core.Tests/ArithmeticTests.cs
@@ -246,7 +246,7 @@ public class ArithmeticTests
     {
         var pair = Def.GetPair((byte)FloatOp.Question);
         Assert.That(pair.EmitOnMatch, Is.True, "Question.EmitOnMatch should be true");
-        Assert.That(pair.EmitOpCode, Is.EqualTo(FloatOp.Select), "Question.EmitOpCode should be Select");
+        Assert.That(pair.EmitOpCode, Is.EqualTo((byte)FloatOp.Select), "Question.EmitOpCode should be Select");
         Assert.That(pair.PairRole, Is.EqualTo(Pair.None), "Question.PairRole should be None");
     }
